+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog Mapgl
 
 ## 2.0.1
+- arrows for edges
+- cut edges at node boundary ports 
 - show head node id in edge tooltip
 - enlarge picked Arc edge